Sensor capture data and data plotter files
once up on the joystick is pressed, the sensor data records data once every ~2 seconds, sends a "blip" across the screen and writes to a CSV with today's date.

the data plotter file attempts to graph out the data collected previously. I still need to work on matplotlib, or possibly pandas to understand how to properly show the information.